+++ tests/unit/test_helpers.py
@@ -1,11 +1,10 @@
 import pytest
-from django import VERSION as django_version
 from django.core.exceptions import ImproperlyConfigured
 from django.test import override_settings
 
 from wagtail_marketing.helpers import SeoHelper
 
-ELLIPSIS_CHARACTER = '…' if django_version >= (2, 2) else '...'
+ELLIPSIS_CHARACTER = '…'
 
 
 class TestPageAdminURLHelper:
@@ -32,19 +31,13 @@ class TestSeoHelper:
         helper = SeoHelper(
             'My mama always said life was like a box of chocolates. You never know what you are gonna get.')
 
-        if django_version >= (2, 2):
-            assert helper.truncated_title == 'My mama always said life was like a box of chocolates. You {}'.format(ELLIPSIS_CHARACTER)
-        else:
-            assert helper.truncated_title == 'My mama always said life was like a box of chocolates. Yo{}'.format(ELLIPSIS_CHARACTER)
-
+        assert helper.truncated_title == f'My mama always said life was like a box of chocolates. You {ELLIPSIS_CHARACTER}'
+    
     @override_settings(WAGTAIL_MARKETING_MAX_TITLE_LENGTH=30)
     def test_title_truncation_with_setting_override(self):
         helper = SeoHelper('Toto, I have a feeling we are not in Kansas anymore.')
 
-        if django_version >= (2, 2):
-            assert helper.truncated_title == 'Toto, I have a feeling we are{}'.format(ELLIPSIS_CHARACTER)
-        else:
-            assert helper.truncated_title == 'Toto, I have a feeling we a{}'.format(ELLIPSIS_CHARACTER)
+        assert helper.truncated_title == f'Toto, I have a feeling we are{ELLIPSIS_CHARACTER}'
 
     @override_settings(WAGTAIL_MARKETING_MAX_TITLE_LENGTH=30)
     def test_seo_title_truncation_with_setting_override(self):
@@ -53,10 +46,7 @@ class TestSeoHelper:
             'True courage is in facing danger when you are afraid',
         )
 
-        if django_version >= (2, 2):
-            assert helper.truncated_title == 'True courage is in facing dan{}'.format(ELLIPSIS_CHARACTER)
-        else:
-            assert helper.truncated_title == 'True courage is in facing d{}'.format(ELLIPSIS_CHARACTER)
+        assert helper.truncated_title == f'True courage is in facing dan{ELLIPSIS_CHARACTER}'
 
     def test_description_truncation_with_the_default_setting(self):
         helper = SeoHelper(
@@ -66,14 +56,9 @@ class TestSeoHelper:
                                'Kansas that you have brains.'
         )
 
-        if django_version >= (2, 2):
-            assert helper.truncated_description == 'If your heads were stuffed with straw, like mine, you would probably ' \
+        assert helper.truncated_description == 'If your heads were stuffed with straw, like mine, you would probably ' \
                                                 'all live in the beautiful places, and then Kansas would have no ' \
-                                                'people at all. I{}'.format(ELLIPSIS_CHARACTER)
-        else:
-            assert helper.truncated_description == 'If your heads were stuffed with straw, like mine, you would probably ' \
-                                                'all live in the beautiful places, and then Kansas would have no ' \
-                                                'people at all.{}'.format(ELLIPSIS_CHARACTER)
+                                                f'people at all. I{ELLIPSIS_CHARACTER}'
 
     @override_settings(WAGTAIL_MARKETING_MAX_DESCRIPTION_LENGTH=30)
     def test_description_truncation_with_setting_override(self):
@@ -82,10 +67,7 @@ class TestSeoHelper:
             search_description='I am content in knowing I am as brave as any best that ever lived, if not braver.'
         )
 
-        if django_version >= (2, 2):
-            assert helper.truncated_description == 'I am content in knowing I am {}'.format(ELLIPSIS_CHARACTER)
-        else:
-            assert helper.truncated_description == 'I am content in knowing I a{}'.format(ELLIPSIS_CHARACTER)
+        assert helper.truncated_description == f'I am content in knowing I am {ELLIPSIS_CHARACTER}'
 
     def test_score_0(self):
         helper = SeoHelper('', search_description='', seo_title='')