+++ tests/urls.py
@@ -2,5 +2,5 @@ from django.urls import include, path
 
 urlpatterns = [
     path('admin/', include('wagtail.admin.urls')),
-    path('', include('wagtail.core.urls')),
+    path('', include('wagtail.urls')),
 ]